feat: first prototype of CLI using do_hydra
Currently implements the variant_annotation and variant_index steps.

Note: During pre-commit pycln introduced a bug in the code by introducing
a functionality that it's only available in python 3.10+.
To fix it we need to revert `path: str | None` to `path: Optional[str]`
in `variant_annotation.py` and `variant_index.py`.