fix(airflow): job args are list of strings
This affects the way that args are provided for the pyspark job specification.
DataprocSubmitJobOperator takes args as a list of strings.
The previous function accepted a list of strings or a dictionary as inputs. When provided a dictionary, it'd parse it into a list of strings.
In https://github.com/opentargets/genetics_etl_python/commit/930155765e7e7c474f877faad0018b329d977e79 I introduced a bug that incorrectly assumed that args were provided as a dict. If a list was provided, formatted_args would remain an empty array.
This is actually the cause for the bug reported here https://github.com/opentargets/genetics_etl_python/pull/201#pullrequestreview-1706949353