ci(artifact): build images into GHCR and GAR
This change introduces building container to the GAR without using long
term credentials (impersonating service account through the OIDC).
From now this action is only run on demand (the ci needs to be triggered
by user) limiting the amount of artifacts that we not utilize.
This action can only be run from tag (not branch)